create a jenkins-button instead of a yui button in makeButton
makeButton is one of the last places in core were YUI is used.
This method is called by a few plugins directly but also when plugins
create inputs of type button with class `submit-button` or `yui-button`.
Frequently they then also attach an onclick handler.

This change will replace the input with a button with class
`jenkins-button` on the fly.
Looking at the code of several plugins that rely on the behaviour it
seems they will not be affected by this change.