[JENKINS-73613] refresh buildhistory widget in all cases (#9624)
* [JENKINS-73592] refresh buildhistory widget when page gets focus

* [JENKINS-73613] refresh buildhistory in all cases

refresh the buildHistory widget instantly when the window gets focus,
e.g. after switching browser tabs but also when one was in another
application and clicks in the browser window.
When the window is visible it will update even when it has no focus,
when the window is hidden no calls to the controller will be done.

Additionally the buildHistory will be updated when it is not on the
first page. So e.g. when someone deletes a build, changes the display
name or the description or when a build is still running the page gets
updated.