[JENKINS-73302] Restore margins around setup wizard alert messages
https://issues.jenkins.io/browse/JENKINS-73302 reports that when a
user of the setup wizard skips the creation of a user and skips the
assignment of the Jenkins URL, the two alert messages no longer have a
margin between them.

This adds a top margin to each of the messages so that when two messages
are displayed, there will be a margin between them.

Testing done:

Checked the visual display with 2 messages and with each message alone.