[JENKINS-73771] do not loose the formField in the response
Implements a crude way to not loose the important part that is the form
field.

This is a quick and dirty approach, a complete fix should also take into
account the describable being bound as in a nested page like
/manage/configure a field name may not even be unique.

However for now this makes the situation less bad without introducing
debt.