Introduce SaveableListener#onDeleted
Usually `Saveable` objects are written, but it can happen on occasion that they get deleted, and it wasn't generating an event for every case.

This provides a more fine-grained event that can be handled by implemented listeners.

In my case, I have a use case in CloudBees CI where I need to clear a cache entry when a Saveable gets deleted from disk.