[JENKINS-73813] Show a notification when scheduling a build fails
When a build is triggered via the UI (either on the job page or via the
buildbuttoncolumn in a view) and the crumb is no longer valid, there was
still the tooltip shown that the build was scheduled, whereas nothing
happened.
This change will show a notification when scheduling failed.