Display appropriate GUI that accurately displays offline by design
Reworked computer icon and associated descriptions to be controlled by the `OfflineCause` implementation. For implementations such as `RetentionStrategy.Demand`, this allows to set an offline cause that can't be interpreted by the end user as an error.

Introduced `pause` and `computer-paused` symbols, used by the on-demand retention strategy to indicate the computer is offline on purpose and is not in error state
Add `getOfflineCause` to `IComputer` and refactor related methods to pull up methods that make sense at the interface level.
Extract `IOfflineCause` interface from `OfflineCause`.