Allow subclasses of UpdateSite to call updateData
The updateData method is required to be called for the
actual update data to be synced. For plugins that may
implement a different method for interacting with an
update site (e.g. an implementation of password or
API key protected update site being planned for the
Managed Update Sites plugin), we need to override
updateDirectlyNow and call updateData with the
retrieved JSON data.