Trim the environment variables keys and values
Add in trimming of the node environment variables key and value values. If you enter values in either fields with leading spaces, trailing spaces or leading and trailing spaces, those spaces would stay and the environment variables would not load. This becomes a real problem when adding `BASE+EXTRA` values, such as `PATH+LOCAL_BIN`.

This change trims the `key` and `value` values.