handle empty and spaces a little differently
After reviewing the help, the last sentence states:

```
If the Value is empty or whitespace-only, it will not be added to the environment, nor will it override or unset any environment variable with the same name that may already exist (e.g. a variable defined by the system).
```

This wasn't true prior to my changes nor any of the commits up to this point. I think this covers all the cases.