Disable `JnlpSlaveRestarterInstallerTest` on Windows CI agents (#10195)
The tests have failed consistently on Windows agents of ci.jenkins.io
since we switched to virtual machines instead of containers.  The tests
fail on the stable-2.479, stable-2.492, and master branches.  The tests
do not fail when run on my Windows 10 and Windows 11 computers in my
home lab.

Needs more investigation, but I would rather not block the 2.492.1
release candidate build for that investigation.