Disable flaky tests on ci.jenkins.io
The tests are known to pass in development environments and sometimes pass
in CI.  However, enough of the tests are flaky on ci.jenkins.io that they
are preventing the delivery of a release candidate for Jenkins 2.492.3.

The same flakiness is visible on the master branch and in pull requests
when the tests are run on ci.jenkins.io.

Jenkins infrastructure help desk issue has been reported as

* https://github.com/jenkins-infra/helpdesk/issues/4604