[JENKINS-75461] add top button for hetero-list, e.g. "Add build step"
Adds a dropdown for adding items to a hetero-list to the top of the list
when there is at least one item.
This is useful when the it is common to have several items in such a
list and the items contain itself configuration.

Enable the flag for job parameters and build steps and post build
actions in freestyle jobs.