Disable WebClient's js in FileParameterValueTest and DirectoryBrowserSupportTest
These two tests looked like frequently failing in the test paths that visit a builds `ws` page, which is having the javascript execution issue
https://github.com/jenkinsci/jenkins-test-harness/issues/946

Here trying to disable the javascript in all the webclient usages in these two tests.
Note, I have disabled in the test methods that didn't visit the `ws` page specifially to keep the `getWebClient()` usage uniform.