[JENKINS-33704] Limit scope of `Jenkins#updateComputerList`
Along the same line as #5882, but for computers

Adding/updating/removing a node to the system was causing all nodes retention strategies to be checked. If you have a lot of nodes, this quickly adds up to an extreme amount of checks.

This change makes the calls local, by only affecting the related nodes.