[JENKINS-68155] Fix state of previous labels when updating a node label
Prior to #5882, trimLabels was global, so any node operation would cause all labels to be rebuilt.

After #5882, the labels to rebuild were taken from the current state of nodes. When updating a node label however, both the old label and the new label needs to be rebuilt.

When a node labels are updated, the previous values are now saved aside in order to compute the proper labels to trim.