[JENKINS-33704] Computer configuration not updated correctly after updating nodes
Several ATHs are failing since #10494

https://ci.jenkins.io/job/Core/job/acceptance-test-harness/job/PR-1956/7/testReport/

This is due to the call to `Jenkins#updateComputers(Node...)` being passed both the old and new object for a node when applying an update. This resulted into the old node configuration to be applied to the computer.
Changed the call to only pass the new node object.

The previous global call to `Jenkins#updateComputerList()` was looking up all nodes attached to the Jenkins node, so was not applying the logic to old node configurations.

Added a unit test catching the regression, as this was only caught in ATH.