[JENKINS-75683] re-introduce actions for Jenkins with custom action.jelly
actions without a URL but a custom action.jelly where not shown
correctly in the header.

This re-introduces them and allows custom rendering of actions again (by
using an action.jelly).

adjust the task taglib so that it renders in the style of the header
when in the header.

Nested tasks work, but when expanded the original looses its tooltip.

move subtasks to the dropdown

use tasks class to fix layout

fix scss formatting