[JENKINS-75683] re-introduce actions for Jenkins with custom action.jelly
adjust the task taglib so that it renders in the style of the header
when in the header.

remove the workaround from the HEader (setting the url) we now correctly
set it to the Jenkins instance for the header (root) actions.

Nested tasks work, but when expanded the original looses its tooltip.

Contains a (non working) experiment of using a jump list for the
embedded actions but for some reason this is not working...

needs investigation.