Fix FileChannelWriterTest and VirtualFileTest by reverting to JUnit 4 (#10698)
* Fix FileChannelWriterTest by reverting to JUnit 4

Test does not fail on my local Windows computer with JUnit 5 or with
JUnit 4.  Reverting this to see if it resolves the issue so that other
pull requests will not be blocked by the failing test.

* Fix VirtualFileTest by reverting back to JUnit 4

(cherry picked from commit dc3ae003fb07aaa280e291e1f7b3de518fde2828)