Distinguish between primary and secondary actions in the header
This PR distinguishes between primary and secondary Actions for the
header.

Primary actions are always shown, and secondary actions are by default
in the Hamburger menu.
When the current page is the page for the secondary action (or a parent
thereof) the action is put in the main header action list and not in the
hamburger menu.