Fix url in search result list when running at root context (#10725)
fix url when running at root

when the rootURL is empty the urls would start with a double slash which
is interpreted by browser like a complete url including domain. That
leads to non working links.
Use the same approach for the url as used by command pallete to include
the additional slash only when the url is not starting with a slash