Restore maven-shade-plugin version setting in cli/pom.xml
Without this change, the build process reports:

```
[WARNING]
[WARNING] Some problems were encountered while building the effective model for org.jenkins-ci.main:cli:jar:2.518-rc37026.a_147d8a_b_56d4
[WARNING] 'build.plugins.plugin.version' for org.apache.maven.plugins:maven-shade-plugin is missing. @ org.jenkins-ci.main:cli:${revision}${changelist}, /home/jenkins/agent/workspace/Core_jenkins_master/cli/pom.xml, line 118, column 15
[WARNING]
[WARNING] It is highly recommended to fix these problems because they threaten the stability of your build.
[WARNING]
[WARNING] For this reason, future Maven versions might no longer support building such malformed projects.
[WARNING]
```

https://ci.jenkins.io/job/Core/job/jenkins/job/master/7510/consoleFull
has examples.

The version was removed in a147d8ab56d4e5dfe7a9729172699e731ef0dcd0 as
part of pull request https://github.com/jenkinsci/jenkins/pull/10788

Testing done

Confirmed that the message is visible before this change and is not
visible after this change

Confirmed that 3.6.0 is the most recent release of the Maven Shade plugin.