Reduce inQueueTaskLookupByAPI flakiness
https://app.launchableinc.com/organizations/jenkins/workspaces/jenkins/data/test-paths/class%3Dhudson.model.QueueTest%23%23%23testcase%3DinQueueTaskLookupByAPI?dateFilter=90d
reports that inQueueTaskLookupByAPI has been flaky in th epast 90 days.
The flakiness seems to be specific to Windows and seems to be related to
the job not be in the queue quickly enough after the call toe schedule
the build.

If the job is not queued, sleep for 3 seonds to allow it time to queue.

Testing done:

Confirmed that the job queues consistently on my Linux computer, so the
sleep statement is not executed.