Reduce flakiness of ComputerStateTest#testUiForConnected
https://app.launchableinc.com/organizations/jenkins/workspaces/jenkins/data/test-paths/class%3Dhudson.cli.ComputerStateTest%23%23%23testcase%3DtestUiForConnected?dateFilter=90d
reports that there have been multiple failures of
ComputerStateTest#testUiForConnected in the last 90 days.  The failures
seem to be related to a race condition that the agent is not yet
disconnected when the state of the agent is checked.  Wait for 1 second
to allow the agent to disconnect before checking agent status.

The average time to run the test is 34 seconds.  This will increase that
average time by 1 second.  Not a significant increase compared to the
benefit of avoiding a flaky test.