Reduce test flakiness (#10809)
* Reduce inQueueTaskLookupByAPI flakiness

https://app.launchableinc.com/organizations/jenkins/workspaces/jenkins/data/test-paths/class%3Dhudson.model.QueueTest%23%23%23testcase%3DinQueueTaskLookupByAPI?dateFilter=90d
reports that inQueueTaskLookupByAPI has been flaky in th epast 90 days.
The flakiness seems to be specific to Windows and seems to be related to
the job not be in the queue quickly enough after the call toe schedule
the build.

If the job is not queued, sleep for 3 seonds to allow it time to queue.

Testing done:

Confirmed that the job queues consistently on my Linux computer, so the
sleep statement is not executed.

* Replace SCSS 1x with 1dppx to silence HTMLUnit warning

Per https://developer.mozilla.org/en-US/docs/Web/CSS/resolution , `x` is
an alias for `dppx`, so let's use `dppx` to avoid warnings from HTMLUnit.

* Reduce flakiness of ComputerStateTest#testUiForConnected

https://app.launchableinc.com/organizations/jenkins/workspaces/jenkins/data/test-paths/class%3Dhudson.cli.ComputerStateTest%23%23%23testcase%3DtestUiForConnected?dateFilter=90d
reports that there have been multiple failures of
ComputerStateTest#testUiForConnected in the last 90 days.  The failures
seem to be related to a race condition that the agent is not yet
disconnected when the state of the agent is checked.  Wait for 1 second
to allow the agent to disconnect before checking agent status.

The average time to run the test is 34 seconds.  This will increase that
average time by 1 second.  Not a significant increase compared to the
benefit of avoiding a flaky test.

* Reduce preventXssInBadgeTooltip flakes with a 2 second sleep

https://app.launchableinc.com/organizations/jenkins/workspaces/jenkins/data/test-paths/class%3Dhudson.model.RunTest%23%23%23testcase%3DpreventXssInBadgeTooltip?dateFilter=90d
shows there have been random test failures of preventXssInBadgeTooltip
in the last 90 days.

Reeduce failures by sleeping before the key assertions in order to allow
Jenkins to complete the initialization of the page.