`Addressable` and `AddressableModelObject` interfaces
These new interfaces provide `getUrl` method, something that was mentioned in `ModelObject` for a long time.

I considered making `Descriptor` implement this new interface, but considering the legacy of existing implementations implementing such methods with different semantic (`JenkinsLocationConfiguration`, maybe others in plugins), I refrained from doing it.

Also considered adding it to `Actionable`. It would be natural because `Action`s implement `getUrlName` and are expected to be addressable under an `Actionable`.

Unfortunately I don't think this is doable without breaking compatibility. A new `Actionable2` could be implemented if this seems really useful.