Remove trivial tooltips from the help link
Before this change, the help link showed a trivial tooltip that
simply repeated the name of the control. Such tooltips distract the user.
Users might even not click the link anymore because they think that the tooltip
already shows the actual help message. Now the additional help text
is available as aria-label only.