Skip runScriptOnOfflineComputer on Windows ci.jenkins.io
Test was slightly flaky before ci.jenkins.io transitioned from Azure to
AWS agents.  Since the transition, the test fails much more frequently.
The failures are not pointing to a real problem, but a change in the
infrastructure.

Infra help desk issue that is investigating:

* https://github.com/jenkins-infra/helpdesk/issues/4730

Let's not waste resources running this test on Windows in CI.

Testing done:

* Ran JenkinsTest#runScriptOnOfflineComputer on 5 different Windows
  computers before making this change.  Execution times ranged from
  21 sec to 45 sec (depending on the age of the Windows computer).
  None of those tests failed.

* Confirmed that the test is skipped on those Windows computers after
  making this change

Those Windows computers are running a mix of Windows 10 and Windows 11
with local SSD disc drives.  They each have at least 16 GB of RAM.