Remove inaccurate javadoc that includes an incorrect link
The Javadoc comment said:

> When executing the resulting Callable, it executes
> FileCallable#act(FileCallable) on this FilePath.

The FileCallable interface does not have an `act(FileCallable)` method.
That absence causes the Javadoc generation to fail on Java 25.

There is a `FilePath.act(FileCallable)` method that could be inserted
instead, but I'm not confident the comment would then be correct.

Testing done:

Confirmed that Javadoc generation fails on Java 25 without this change.

Confirmed that Javadoc generation passes on Java 25 with this change.