Include URL in loadJSON and loadJSONHTML exception messages
Includes the basename of the original exception and the message of the
original exception so that the exception message retains the information
from the original exception.

Without the original exception message and exception class, it can be
difficult to understand the cause of the message.  Was it a connection
timeout or a connection reset or a hostname not found or something else?