Fix NPE in ReverseBuildTrigger due to null upstream threshold (#23916)
* Fix NPE in ReverseBuildTrigger due to null upstream threshold (JENKINS-39044)

* Use ReverseTrigger constructor that is not deprecated

Has the same result as the deprecated constructor that was used before.

* Add automated test of JENKINS-39044

If the ReverseTrigger threshold were null, a null pointer exception would
be thrown.  The config round trip operation resets a null threshold to
SUCCESS, so this test reaches inside the job definition and changes the
threshold directly.

Testing done:

* Confirmed that the modified automated test fails when the change to
  ReverseTrigger is removed.
* Confirmed that the modified automated test passes when the change to
  ReverseTrigger is included.

* Link to GitHub issue

---------

Co-authored-by: Mark Waite <mark.earl.waite@gmail.com>