Re-add `out` id to Console Output (#25953)
Partially reverts 331c768.

In #8960 the `out` id was added to `pre.console-output`'s  "otherwise"
case in `console.jelly`.

In 331c768, the `out` id was removed from `otherwise` when console output
was moved from `console.jelly` to `console-log.jelly`.

This may have broken the "Copy" button in some instances.

When clicked, the copy button will copy the inner text of a container
with the id referred to by the `ref` attribute of the copy button. The
copy button's ref attribute is `out`.

Wikimedia's Jenkins has no element with the `out` id on the console
page, which leads to copying `null` to the clipboard.

This was filed as https://phabricator.wikimedia.org/T410633.

Co-authored-by: Mark Waite <mark.earl.waite@gmail.com>